refactored folder ui and badges
