+++ src/components/badges/CategoryBadge.vue
